# tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "actjs",
    "description": "Self-hosted actor framework HTTP API",
    "version": "0.3.0"
  },
...
  "paths": {
...
    "/v1/manifest": {
      "get": {
        "summary": "Resolve a fresh manifest",
        "tags": [
          "manifest"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "toJSONSchema"
          },
          {
            "schema": {
              "type": "object",
              "shape": {
                "root": {
                  "def": {
                    "type": "optional",
                    "innerType": {
                      "def": {
                        "type": "union",
                        "options": [
                          {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          },
                          {
                            "def": {
                              "type": "array",
                              "element": {
                                "def": {
                                  "type": "string"
                                },
                                "type": "string",
                                "format": null,
                                "minLength": null,
                                "maxLength": null
                              }
                            },
                            "type": "array",
                            "element": {
                              "def": {
                                "type": "string"
                              },
                              "type": "string",
                              "format": null,
                              "minLength": null,
                              "maxLength": null
                            }
                          }
                        ]
                      },
                      "type": "union",
                      "options": [
                        {
                          "def": {
                            "type": "string"
                          },
                          "type": "string",
                          "format": null,
                          "minLength": null,
                          "maxLength": null
                        },
                        {
                          "def": {
                            "type": "array",
                            "element": {
                              "def": {
                                "type": "string"
                              },
                              "type": "string",
                              "format": null,
                              "minLength": null,
                              "maxLength": null
                            }
                          },
                          "type": "array",
                          "element": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          }
                        }
                      ]
                    }
                  },
                  "type": "optional"
                },
                "dep": {
                  "def": {
                    "type": "optional",
                    "innerType": {
                      "def": {
                        "type": "union",
                        "options": [
                          {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          },
                          {
                            "def": {
                              "type": "array",
                              "element": {
                                "def": {
                                  "type": "string"
                                },
                                "type": "string",
                                "format": null,
                                "minLength": null,
                                "maxLength": null
                              }
                            },
                            "type": "array",
                            "element": {
                              "def": {
                                "type": "string"
                              },
                              "type": "string",
                              "format": null,
                              "minLength": null,
                              "maxLength": null
                            }
                          }
                        ]
                      },
                      "type": "union",
                      "options": [
                        {
                          "def": {
                            "type": "string"
                          },
                          "type": "string",
                          "format": null,
                          "minLength": null,
                          "maxLength": null
                        },
                        {
                          "def": {
                            "type": "array",
                            "element": {
                              "def": {
                                "type": "string"
                              },
                              "type": "string",
                              "format": null,
                              "minLength": null,
                              "maxLength": null
                            }
                          },
                          "type": "array",
                          "element": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          }
                        }
                      ]
                    }
                  },
                  "type": "optional"
                }
              }
            },
            "in": "query",
            "name": "def"
          },
          {
            "schema": "object",
            "in": "query",
            "name": "type"
          },
          {
            "in": "query",
            "name": "parse"
          },
          {
            "in": "query",
            "name": "safeParse"
          },
          {
            "in": "query",
            "name": "parseAsync"
          },
          {
            "in": "query",
            "name": "safeParseAsync"
          },
          {
            "in": "query",
            "name": "spa"
          },
          {
            "in": "query",
            "name": "encode"
          },
          {
            "in": "query",
            "name": "decode"
          },
          {
            "in": "query",
            "name": "encodeAsync"
          },
          {
            "in": "query",
            "name": "decodeAsync"
          },
          {
            "in": "query",
            "name": "safeEncode"
          },
          {
            "in": "query",
            "name": "safeDecode"
          },
          {
            "in": "query",
            "name": "safeEncodeAsync"
          },
          {
            "in": "query",
            "name": "safeDecodeAsync"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "def": {
                    "type": "object",
                    "shape": {
                      "sha256": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "resolved": {
                        "def": {
                          "type": "record",
                          "keyType": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          },
                          "valueType": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          }
                        },
                        "type": "record",
                        "keyType": {
                          "def": {
                            "type": "string"
                          },
                          "type": "string",
                          "format": null,
                          "minLength": null,
                          "maxLength": null
                        },
                        "valueType": {
                          "def": {
                            "type": "string"
                          },
                          "type": "string",
                          "format": null,
                          "minLength": null,
                          "maxLength": null
                        }
                      },
                      "constraints": {
                        "def": {
                          "type": "record",
                          "keyType": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          },
                          "valueType": {
                            "def": {
                              "type": "array",
                              "element": {
                                "def": {
                                  "type": "object",
                                  "shape": {
                                    "range": {
                                      "def": {
                                        "type": "string"
                                      },
                                      "type": "string",
                                      "format": null,
                                      "minLength": null,
                                      "maxLength": null
                                    },
                                    "path": {
                                      "def": {
                                        "type": "array",
                                        "element": {
                                          "def": {
                                            "type": "string"
                                          },
                                          "type": "string",
                                          "format": null,
                                          "minLength": null,
                                          "maxLength": null
                                        }
                                      },
                                      "type": "array",
                                      "element": {
                                        "def": {
                                          "type": "string"
                                        },
                                        "type": "string",
                                        "format": null,
                                        "minLength": null,
                                        "maxLength": null
                                      }
                                    }
                                  }
                                },
                                "type": "object"
                              }
                            },
                            "type": "array",
                            "element": {
                              "def": {
                                "type": "object",
                                "shape": {
                                  "range": {
                                    "def": {
                                      "type": "string"
                                    },
                                    "type": "string",
                                    "format": null,
                                    "minLength": null,
                                    "maxLength": null
                                  },
                                  "path": {
                                    "def": {
                                      "type": "array",
                                      "element": {
                                        "def": {
                                          "type": "string"
                                        },
                                        "type": "string",
                                        "format": null,
                                        "minLength": null,
                                        "maxLength": null
                                      }
                                    },
                                    "type": "array",
                                    "element": {
                                      "def": {
                                        "type": "string"
                                      },
                                      "type": "string",
                                      "format": null,
                                      "minLength": null,
                                      "maxLength": null
                                    }
                                  }
                                }
                              },
                              "type": "object"
                            }
                          }
                        },
                        "type": "record",
                        "keyType": {
                          "def": {
                            "type": "string"
                          },
                          "type": "string",
                          "format": null,
                          "minLength": null,
                          "maxLength": null
                        },
                        "valueType": {
                          "def": {
                            "type": "array",
                            "element": {
                              "def": {
                                "type": "object",
                                "shape": {
                                  "range": {
                                    "def": {
                                      "type": "string"
                                    },
                                    "type": "string",
                                    "format": null,
                                    "minLength": null,
                                    "maxLength": null
                                  },
                                  "path": {
                                    "def": {
                                      "type": "array",
                                      "element": {
                                        "def": {
                                          "type": "string"
                                        },
                                        "type": "string",
                                        "format": null,
                                        "minLength": null,
                                        "maxLength": null
                                      }
                                    },
                                    "type": "array",
                                    "element": {
                                      "def": {
                                        "type": "string"
                                      },
                                      "type": "string",
                                      "format": null,
                                      "minLength": null,
                                      "maxLength": null
                                    }
                                  }
                                }
                              },
                              "type": "object"
                            }
                          },
                          "type": "array",
                          "element": {
                            "def": {
                              "type": "object",
                              "shape": {
                                "range": {
                                  "def": {
                                    "type": "string"
                                  },
                                  "type": "string",
                                  "format": null,
                                  "minLength": null,
                                  "maxLength": null
                                },
                                "path": {
                                  "def": {
                                    "type": "array",
                                    "element": {
                                      "def": {
                                        "type": "string"
                                      },
                                      "type": "string",
                                      "format": null,
                                      "minLength": null,
                                      "maxLength": null
                                    }
                                  },
                                  "type": "array",
                                  "element": {
                                    "def": {
                                      "type": "string"
                                    },
                                    "type": "string",
                                    "format": null,
                                    "minLength": null,
                                    "maxLength": null
                                  }
                                }
                              }
                            },
                            "type": "object"
                          }
                        }
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/v1/manifest/{sha}": {
      "get": {
        "summary": "Retrieve a stored manifest",
        "tags": [
          "manifest"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "toJSONSchema",
            "required": true
          },
          {
            "schema": {
              "type": "object",
              "shape": {
                "sha": {
                  "def": {
                    "type": "string",
                    "checks": [
                      {}
                    ]
                  },
                  "type": "string",
                  "format": "regex",
                  "minLength": null,
                  "maxLength": null
                }
              }
            },
            "in": "path",
            "name": "def",
            "required": true
          },
          {
            "schema": "object",
            "in": "path",
            "name": "type",
            "required": true
          },
          {
            "in": "path",
            "name": "parse",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParse",
            "required": true
          },
          {
            "in": "path",
            "name": "parseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeParseAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "spa",
            "required": true
          },
          {
            "in": "path",
            "name": "encode",
            "required": true
          },
          {
            "in": "path",
            "name": "decode",
            "required": true
          },
          {
            "in": "path",
            "name": "encodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "decodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecode",
            "required": true
          },
          {
            "in": "path",
            "name": "safeEncodeAsync",
            "required": true
          },
          {
            "in": "path",
            "name": "safeDecodeAsync",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "def": {
                    "type": "object",
                    "shape": {
                      "sha256": {
                        "def": {
                          "type": "string"
                        },
                        "type": "string",
                        "format": null,
                        "minLength": null,
                        "maxLength": null
                      },
                      "resolved": {
                        "def": {
                          "type": "record",
                          "keyType": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          },
                          "valueType": {
                            "def": {
                              "type": "string"
                            },
                            "type": "string",
                            "format": null,
                            "minLength": null,
                            "maxLength": null
                          }
                        },
                        "type": "record",
                        "keyType": {
                          "def": {
                            "type": "string"
                          },
                          "type": "string",
                          "format": null,
                          "minLength": null,
                          "maxLength": null
                        },
                        "valueType": {
                          "def": {
                            "type": "string"
                          },
                          "type": "string",
                          "format": null,
                          "minLength": null,
                          "maxLength": null
                        }
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}